Run rustfmt over entire repository
Because this is a monorepo with multiple Cargo.toml, cargo +nightly fmt doesn't format sub-crates. (If there is functionality to do so, I haven't found it, and it should likely be added.)

In order to format a subcrate you must cd to the directory which contains the Cargo.toml describing the crate you are working on and then run rustfmt.