Reconsider macro Long_Name in favor of "Display Name".
Bidi_Class uses names with spaces and hyphens, meaning that the naive $ident match will not work. This also cuts down on chaff, because there are no longer two matches in the macro or repeated information (other than that in the docs).

The doc repetition I think is valid, as docs should be considered separately and just populating them with the display name would be redundant.