Character Map trait.
This is where the transformation to output formats will live.

This trait may be altered slightly, such that T does not have to impl Display (which is currently what is used to output it to the serialized format) but rather to have a provided Fn(T) -> impl Display on format fn to present the item for serialization.

Signed-off-by: CAD97 <cad97@cad97.com>