un lazy_static UnicodeData
When compiled with optimizations, this is still near instant.
It takes some time without optimizations,
but this isn't an operation that I expect anyone to be running often.

This also avoids the problem of leaking the memory used for the info.

In any case, we can and should be passing around these shared values
rather than relying on lazy_static and globals.

Signed-off-by: CAD97 <cad97@cad97.com>