[ucd/category] Add all_values() static fn
The `all_values()` static function is a new API for character
properties implemented as *flat* `enum` types in Rust, like some of
Unicode properties of type *Catalog* or *Enumeration*, to return an
exhaustive list of all possible values.

The function is useful mainly in writing integration tests, but since is
efficient in implementation, can be used in application logic, as well.

The current list is manually created and maintained. With migration to
`char_property` macros, the list will be auto-generated. (Possibly
stored in a private module and only exposed via `all_values()` API.)

See <https://github.com/behnam/rust-unic/issues/66>