Why is this faster? But hey, no more unsafe char around!
This makes it take ~0.98ms (on my machine) to iterate all forwards
as opposed to ~1.2ms (on my machine).

Forward and Reverse iteration are now within margin of error
of each other, which means (hopefully) neither are getting an
optimization which is being denied to the other.

I still have no concrete idea why the reverse->filter_map
is ~20% faster than just the filter_map. My only guess is branch
prediction, which means it would be very CPU dependent.