[char/property] Add CharProperty trait
Add `CharProperty` trait as the core main contract for Character
Property definition, which in return requires defining the domain in
which the property is defined.

We should expand `char_property!` macro to accept these values and
implement the trait internally.

Also, we can see if we want to also enforce choosing a range for the
property. At this moment, `Age` yet doesn't have a contract trait.