[char/property] Add long_name()/human_name() to EnumeratedCharProperty
Add `long_name(&self)`/`human_name(&self)` to the
`EnumeratedCharProperty` contract and add values to the implementing
types.

Also needs to be implemented by `char_property!` macro.

These name methods *may* not belong to *Enumerated* contract, but this
is the only contract we have right now that provides them, so let's
start with having them here.

Bascially, looks like we have to major categories of property range
types: those with pre-defined set of possible values (*Enumerated*, and
possibly other ones), and those with numeric/complex values, which don't
have *named* variants for all of them, which would result in
name-generation to be runtime and not compile time.

For the second group, we can add another parent contract to implement
these methods as `fn *_name(&self) -> String`, which are clearly more
expensive and may not be used when `&'static str` is accepted.

This way, we keep the API similar enough to remember, but also not
enforcing runtime overhead.