[ucd] Drop `use ...::*` lines outside of data or tests
the `use ...::*` is considered a bad practice and should be avoided as
much as possible, or limited to the scope where it's needed.

A major case we have needed them in UCD components is when loading data
from `rsv` files, where we don't want to prefix values in every single
line with the type name. To limit the scope of the wildcard usage, I
have moved these to a local `mod data`.