[ucd/age] Change Age type to struct
The *Age* property was implemented as an `enum` type which was closely
mirroring the core `Option` type. Now that we have a clear definition of
`Partial` properties, it makes sense to update `Age` to be a Custom
property type (`struct`) with *actual* `UnicodeVersion` value.

But, there was a need to update one thing about the Age, it's ordering!
Previously, Age was ordered in the same direction as `UnicodeVersion`,
and `Unassigned` value was larger than any `Assigned(UnicodeVersion)`.
That didn't make much sense in common understanding of *age*, but was
consistent with `UnicodeVersion`.

Now, because of the fact that we cannot change the ordering of `Option`
variants (`None` and `Some(UnicodeVersion)`, in this case), and also to
put the meaning of comparison closer to the common understanding of the
word, we reverse the direciton of ordering for `Age`. Therefore, `Age`
value for character `'A'` is *greater than or equal to* any other
character in Unicode, assigned or unassigned.