[gen/download] Simplify the config format and download script
There's no more code in the download module for each source, and the
list of source comes only from the config file, as well as the file
paths.

For each source, two types of data is supported: normal (or main)
data, and test data. The former goes into `/data/<source>/<name>`, while
the latter goes into `/data/<source>/test/<name>`.

The config file is also simplified to not have any duplicate content as
much as possible.

One major pattern to follow is that each data source has a *base URL*,
which contains the domain name and path to a specific directory for the
version. Nothing goes wrong in the code if this is not maintained, but
it's good practice to keep it so, and create new source config tables if
the domain or path-to-version are different.