[gen] Split generate module into reader and writer
Generating data files useful for Rust code from data sources is not
always a one-to-one mapping. First, we have the fact that some sources,
like `UnicodeData.txt` provides data for many different UNIC components.
Second, we have components that depend on multiple sources. We don't
want to duplicate any code for processing the data, and don't want to
make random cross-file dependencies. So, it's better to split the
modules into two parts: reading source files, and write into component
table files.

The `reader` module provides access to raw data files (from `/data/`)
and is only responsible of parsing the file and returning the value(s)
we're interested in. For example, `reader::ucd::readme` reads
`/data/ucd/ReadMe.txt` file and provides `UNICODE_VERSION`, which is
extracted data from the file content. Most files only provide one big
data table.

All data reading in `reader` is performed lazily, via `lazy_static`,
freeing caller from loading data files explictly and passing around
instances.

The `writer` module is responsible for generating `tables` files for
`unic` components. For example, `writer::ucd::age` is responsible for
creating files under `/unic/ucd/age/tables/`, which depends on
`reader::ucd::readme` (for `UNICODE_VERSION`) and
`reader::ucd::derived_age` (for `AGE_DATA`).

Other Notes:

We were using `std::io::Result` for our error-management, which is not
flexible enough for this case, as different kinds of `Error`s need to be
supported. Because it was a half-baked solution, I just dropped all of
it and replaced error-handling with `unwrap()` calls. We can (and
should) fix this later by using `error-chain` crate and define a clear
error-handling process. Until then, `unwrap()` and `expect()` work good
enough.

Also simplified many fn signatures with dropping the generics and switch
to concrete types, like `PathBuf`, which is good enough for a scripting
enviroment, when we know that we're only passing in one type of value.
We can change this in the future, but only if it really is needed.
Otherwise, it's just extra complexity for a simple task.