[ucd] Add new segment component
Add `unic-ucd-segment` component with initial implementation of three
main segmentation-related properties:
* `Grapheme_Cluster_Break`,
* `Word_Break`, and
* `Sentence_Break`.

Current implementation uses `char_property!()` macro for
`EnumeratedCharProperty` implementation, which only supports
`TotalCharProperty`.

Since the `Other` (abbr: `XX`) value in all these properties are notions
of non-existance of breaking property, we want to switch to
`PartialCharProperty` domain type and use `Option<enum>`. This is left
as a separate step because it needs changes to the macro.