[char/property] Update syntax for Enumerated mode
Because of a bug in Rust compiler before `1.20.0` release, the old
sytanx won't allow having multi-line doc-blocks on enum variants. Since
we don't want to update our minimum-supported-rustc-version at the
moment, need to update the macro syntax to mitigate the bug.

The new syntax is loosly based on the `match` syntax, plus the newly
accepted RFC to allow extra `VERTICAL LINE` prefixes.