Add new unic-ucd-common component
There are some core properties that are commonly used in Unicode
algorithsm, as well as in applications directly, and are not specific to
any single area. Examples of these properties are `Alphabetic` and
`White_Space`.

Also, there are some *resolved* properties that are used commonly, like
*Numeric* and *Alphanumeric*, which are commonly defined as based on
*General_Category* and *Alphabetic* properties. Since they are common in
applications, it makes sense to provided optimized implementations.

This new componet, `unic-ucd-common`, hosts these properties.