[ucd/segment] Make all properties Total
Turns out using these properties is easier as a total property, instead
of partial, as we implemented initially. The reason for it is the fact
that the default value (called `Other` in this context) is usually used
in the same level as other values.

Update variant name for `WordBreak::DoubleQuote`, which was using the
abbr name instead.

Add missing component version test against `unic-ucd-core`.