[data/ucd] Revert GraphemeBreakTest.txt, WordBreakTest.txt
The algorithm has changed, dropping some property values and defining
new ones.

The values used in this version of the tests files do not match the
values defined in the Property Value Aliasas file, therefore resulting
in errors.

In addition, we need to update the Unicode Segmentation Algorithm
implementation before we can run these tests.