[*/lib.rs] Switch back to warn for everything, except `unsafe_code`
Looks like Cargo has issues with silenting lint warnings on external
dependencies (https://github.com/rust-lang/cargo/issues/5998), so until
that's fixed, let's go back to warning for everything that can sudenly
change.

From the looks of it, the only lint rule that won't be affected by
external libraries or the copiler out of a sudden is `unsafe_code`.
Therefore, holding on to that one.